fix(set-min-max-customtimerange):
Sets min & max dates for max & min date picker (respectively). This means that users cannot select a start date that is after the end date, and users cannot select an end date that is after the start date.
Caveat to all of this is that these limitations are not respected when it comes to setting the time. This is an issue with the dependency that I filed here:

https://github.com/Hacker0x01/react-datepicker/issues/2184